requirements.txt
catkin-pkg==0.4.13
loguru==0.3.2
matplotlib==3.1.1
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==6.1.0
PyYAML==5.1.2
requests==2.22.0
rospkg==1.1.10
tensorflow-gpu==1.12.0
utm==0.5.0